Limitation for Long type finally removed: support for integer types
### What's done:
- removed limitation for integers to be long
- TOML spec: Arbitrary 64-bit signed integers (from −2^63 to 2^63−1) should be accepted and handled losslessly. If an integer cannot be represented losslessly, an error must be thrown
- so right now we support Byte, Short and Int. But if there is an overflow, we throw an exception

+++ ktoml-core/src/commonTest/kotlin/com/akuleshov7/ktoml/decoders/ReadMeExampleTest.kt
@@ -22,10 +22,11 @@ class ReadMeExampleTest {
     data class Table1(
         // nullable values, from toml you can pass null/nil/empty value to this kind of a field
         val property1: Long?,
-        // please note, that according to the specification of toml integer values should be represented with Long
-        val property2: Long,
-        // no need to pass this value as it has the default value and is NOT REQUIRED
-        val property3: Long = 5
+        // please note, that according to the specification of toml integer values should be represented with Long,
+        // but we allow to use Int/Short/etc. Just be careful with overflow
+        val property2: Int,
+        // no need to pass this value in the input as it has the default value and so it is NOT REQUIRED
+        val property3: Short = 5
     )
 
     @Serializable
